uid类型改为64

+++ gameServer/center/internal/handler.go
@@ -26,7 +26,7 @@ func init() {
 //玩家在本服节点登录
 func SelfNodeAddPlayer(args []interface{}) {
 	log.Debug("at SelfNodeAddPlayer %v", args)
-	uid := args[0].(int)
+	uid := args[0].(int64)
 	ch := args[1].(*chanrpc.Server)
 	Users[uid] = ch
 	cluster.Broadcast(cost.GamePrefix, "NotifyOtherNodeLogin", uid, conf.ServerName())
@@ -35,7 +35,7 @@ func SelfNodeAddPlayer(args []interface{}) {
 //本服玩家登出
 func SelfNodeDelPlayer(args []interface{}) {
 	log.Debug("at SelfNodeDelPlayer %v", args)
-	uid := args[0].(int)
+	uid := args[0].(int64)
 	delete(Users, uid)
 	cluster.Broadcast(cost.GamePrefix, "NotifyOtherNodelogout", uid)
 }
@@ -43,7 +43,7 @@ func SelfNodeDelPlayer(args []interface{}) {
 //玩家在别的节点登录了
 func NotifyOtherNodeLogin(args []interface{}) {
 	log.Debug("at NotifyOtherNodeLogin %v", args)
-	uid := args[0].(int)
+	uid := args[0].(int64)
 	ServerName := args[1].(string)
 	OtherUsers[uid] = ServerName
 }
@@ -51,13 +51,13 @@ func NotifyOtherNodeLogin(args []interface{}) {
 //玩家在别的节点登出了
 func NotifyOtherNodelogout(args []interface{}) {
 	log.Debug("at NotifyOtherNodelogout %v", args)
-	uid := args[0].(int)
+	uid := args[0].(int64)
 	delete(OtherUsers, uid)
 }
 
 //发消息给别的玩家
 func GoMsgToUser(args []interface{}) {
-	uid := args[0].(int)
+	uid := args[0].(int64)
 	FuncName := args[1].(string)
 	ch, ok := Users[uid]
 	if ok {
